+++ frontend/src/pages/homePage/desktop/utils/calculateTimeToPositionUtil.ts
