+++ frontend/src/pages/homePage/desktop/utils/sortWorkBlocksUtil.ts
@@ -1,6 +1,6 @@
 import type { WorkBlockType } from '@/types/workCard.type';
 import dayjs from 'dayjs';
-import calculateTimeToPosition from './calculateTimeToPosition';
+import calculateTimeToPosition from './calculateTimeToPositionUtil';
 import isTimeOverlapping from '@/lib/dnd/utils/isTimeOverlappingUtil';
 import { getYCoordinate } from '@/constants/workTimeCoordinate';
 import type { CropNameType } from '@/types/crop.type';
